perf(build): prune transitive heavy includes
- use json_fwd.hpp and move json
  formatting to .cpp helper
- drop unused <fmt/format.h>
- move <fmt/std.h> to the .cpp files that
  actually format std types
- split fmt-dependent formatter into
  enum_formatter.h so config.h no longer pulls fmt

Frontend parse time: 162.9s → 142.3s (-12.6%)
