+++ src/linyaps_box/log/macro.h
@@ -7,7 +7,6 @@
 #include "linyaps_box/log/utils.h"
 
 #include <fmt/format.h>
-#include <fmt/std.h>
 
 #include <string_view>
